# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 山东大学黑客松 API
  version: 0.0.1
  description: |
    根据仓库中的 Spring MVC Controller、DTO/VO、校验注解、鉴权拦截器与业务状态码整理。
    除 SDU Pass 回调外，业务接口统一返回 `{code, data, msg}`；成功时 `code` 为 200。
    受保护接口使用访问令牌，服务端同时兼容 `Authorization: Bearer <token>` 与直接传入令牌，本文档按标准 Bearer 方式描述。
    `x-required-roles` 和 `x-required-window` 是对仓库 `@Require` 约束的补充说明。
servers:
  - url: http://localhost:8080
    description: 本地开发环境
...
paths:
...
  /authority/event/{eventId}:
    parameters:
      - {$ref: '#/components/parameters/EventId'}
    get:
      tags: [权限]
      operationId: getAuthorityListByEvent
      summary: 查询赛事相关人员
      security: [{}, {bearerAuth: []}]
      responses:
        '200': {description: 获取成功, content: {application/json: {schema: {$ref: '#/components/schemas/AuthorityUserListResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
...
  /event:
    post:
      tags: [赛事]
      operationId: createEvent
      summary: 创建赛事
      security: [{bearerAuth: []}]
      x-required-roles: [SUPER]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CreateEventDTO'}}}
      responses:
        '200': {description: 创建成功, content: {application/json: {schema: {$ref: '#/components/schemas/EventIdResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /event/{eventId}:
    parameters:
      - {$ref: '#/components/parameters/EventId'}
    get:
      tags: [赛事]
      operationId: getEvent
      summary: 获取赛事详情及赛道名录
      security: [{}, {bearerAuth: []}]
      responses:
        '200': {description: 获取成功, content: {application/json: {schema: {$ref: '#/components/schemas/EventInfoResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
    put:
      tags: [赛事]
      operationId: updateEvent
      summary: 更新赛事
      security: [{bearerAuth: []}]
      x-required-roles: [EVENT_ADMIN]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UpdateEventDTO'}}}
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
    delete:
      tags: [赛事]
      operationId: deleteEvent
      summary: 删除赛事
      security: [{bearerAuth: []}]
      x-required-roles: [SUPER]
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /event/{eventId}/track:
    parameters:
      - {$ref: '#/components/parameters/EventId'}
    post:
      tags: [赛事, 赛道]
      operationId: createTrack
      summary: 在赛事下创建赛道
      security: [{bearerAuth: []}]
      x-required-roles: [EVENT_ADMIN]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CreateTrackDTO'}}}
      responses:
        '200': {description: 创建成功, content: {application/json: {schema: {$ref: '#/components/schemas/TrackIdResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /event/list:
    post:
      tags: [赛事]
      operationId: getEventPage
      summary: 分页查询赛事
      security: [{}, {bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/QueryEventDTO'}}}
      responses:
        '200': {description: 获取成功, content: {application/json: {schema: {$ref: '#/components/schemas/EventPageResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
...
  /event/{eventId}/team-config:
    put:
      tags: [队伍]
      operationId: updateEventTeamConfig
      summary: 配置赛事队伍人数上下限
      security: [{bearerAuth: []}]
      x-required-roles: [EVENT_ADMIN]
      parameters: [{$ref: '#/components/parameters/EventId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UpdateTeamConfigDTO'}}}
      responses:
        '200': {$ref: '#/components/responses/OkResponse'}
        default: {$ref: '#/components/responses/ErrorResponse'}
...
  /file/event/{eventId}:
    parameters:
      - {$ref: '#/components/parameters/EventId'}
    post:
      tags: [文件]
      operationId: presignEventAssetUpload
      summary: 签发赛事资源上传凭证
      security: [{bearerAuth: []}]
      x-required-roles: [EVENT_ADMIN]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UploadDTO'}}}
      responses:
        '200': {description: 已签发上传凭证, content: {application/json: {schema: {$ref: '#/components/schemas/UploadResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录或兑换后返回的 ACCESS 类型 JWT。
  parameters:
    EventId: {name: eventId, in: path, required: true, schema: {type: integer, format: int32, minimum: 1}, description: 赛事 ID}
...
    Page: {name: page, in: query, required: false, schema: {type: integer, minimum: 1, default: 1}, description: 页码，从 1 开始}
    Size: {name: size, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 10}, description: 每页条数}
  responses:
    OkResponse:
      description: 操作成功
      content:
        application/json:
          schema: {$ref: '#/components/schemas/VoidResult'}
    ErrorResponse:
      description: 业务异常、鉴权失败、参数错误或系统异常
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResult'}
          examples:
            parameterError: {summary: 参数错误, value: {code: 4000, data: null, msg: 参数错误}}
            unauthorized: {summary: 未携带令牌, value: {code: 1004, data: null, msg: 未携带token}}
            notFound: {summary: 资源不存在, value: {code: 3001, data: null, msg: 赛事不存在}}
  schemas:
    LocalDateTime:
      type: string
      description: ISO-8601 本地日期时间，不包含时区偏移；服务端业务时区为 Asia/Shanghai。
      pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d+)?$'
      example: '2026-08-30T14:30:00'
    EventStatus: {type: string, enum: [赛前准备, 报名阶段, 赛前休整, 赛事进行, 赛事结束]}
...
    AuthorityType: {type: string, enum: [超级管理员, 赛事管理员, 赛事评委]}
    FileScope:
      type: string
      enum: [AVATAR, EVENT_COVER, TRACK_ATTACHMENT, SUBMIT_ARCHIVE, SUBMIT_VIDEO, SUBMIT_DOC, SUBMIT_IMAGE, MILESTONE, APPEAL, CERT_TEMPLATE, CERT_SEAL, CERTIFICATE, EXPORT, SHOWCASE]
...
    CreateEventDTO:
      type: object
      required: [name, regBeg, regEnd, liveBeg, liveEnd]
      properties:
        name: {type: string, minLength: 1}
        regBeg: {$ref: '#/components/schemas/LocalDateTime'}
        regEnd: {$ref: '#/components/schemas/LocalDateTime'}
        liveBeg: {$ref: '#/components/schemas/LocalDateTime'}
        liveEnd: {$ref: '#/components/schemas/LocalDateTime'}
        introduction: {type: string}
        tags: {type: string, description: 仓库当前以字符串保存的赛事标签}
    UpdateEventDTO:
      type: object
      required: [name, regBeg, regEnd, liveBeg, liveEnd, version]
      properties:
        name: {type: string, minLength: 1}
        regBeg: {$ref: '#/components/schemas/LocalDateTime'}
        regEnd: {$ref: '#/components/schemas/LocalDateTime'}
        liveBeg: {$ref: '#/components/schemas/LocalDateTime'}
        liveEnd: {$ref: '#/components/schemas/LocalDateTime'}
        introduction: {type: string}
        tags: {type: string}
        notice: {type: string}
        version: {type: integer, format: int32, description: 乐观锁版本号}
    QueryEventDTO:
      type: object
      properties:
        name: {type: string}
        statuses: {type: array, items: {$ref: '#/components/schemas/EventStatus'}}
        tags: {type: array, items: {type: string}}
    CreateTrackDTO:
      type: object
      required: [name, descMd]
      properties:
        name: {type: string, minLength: 1}
        descMd: {type: string, minLength: 1, description: Markdown 格式赛道描述}
...
    UpdateTeamConfigDTO:
      type: object
      required: [minSize, maxSize, version]
      properties:
        minSize: {type: integer, minimum: 1}
        maxSize: {type: integer, minimum: 1, maximum: 100}
        version: {type: integer, format: int32}
...
    UploadDTO:
      type: object
      required: [scope, filename, sizeBytes]
      properties:
        scope: {$ref: '#/components/schemas/FileScope'}
        filename: {type: string, minLength: 1, maxLength: 200}
        sizeBytes: {type: integer, format: int64, minimum: 1}
...
    AuthorityUserVO:
      type: object
      properties:
        userId: {type: integer, format: int32}
        name: {type: string}
        type: {$ref: '#/components/schemas/AuthorityType'}
...
    EventIdVO: {type: object, properties: {eventId: {type: integer, format: int32}}}
    TrackIdVO: {type: object, properties: {trackId: {type: integer, format: int32}}}
...
    EventBriefVO:
      type: object
      properties:
        eventId: {type: integer, format: int32}
        name: {type: string}
        status: {$ref: '#/components/schemas/EventStatus'}
        regBeg: {$ref: '#/components/schemas/LocalDateTime'}
        regEnd: {$ref: '#/components/schemas/LocalDateTime'}
        liveBeg: {$ref: '#/components/schemas/LocalDateTime'}
        liveEnd: {$ref: '#/components/schemas/LocalDateTime'}
        tags: {type: [string, 'null']}
    TrackBriefVO:
      type: object
      properties:
        trackId: {type: integer, format: int32}
        name: {type: string}
    EventInfoVO:
      type: object
      properties:
        eventId: {type: integer, format: int32}
        name: {type: string}
        status: {$ref: '#/components/schemas/EventStatus'}
        regBeg: {$ref: '#/components/schemas/LocalDateTime'}
        regEnd: {$ref: '#/components/schemas/LocalDateTime'}
        liveBeg: {$ref: '#/components/schemas/LocalDateTime'}
        liveEnd: {$ref: '#/components/schemas/LocalDateTime'}
        introduction: {type: [string, 'null']}
        tags: {type: [string, 'null']}
        notice: {type: [string, 'null']}
        teamMinSize: {type: integer, minimum: 1}
        teamMaxSize: {type: integer, minimum: 1}
        tracks: {type: array, items: {$ref: '#/components/schemas/TrackBriefVO'}}
        version: {type: integer, format: int32}
...
    UploadVO:
      type: object
      properties:
        fileId: {type: integer, format: int64}
        uploadUrl: {type: string, format: uri, description: 对象存储预签名 PUT 地址}
        contentType: {type: string}
        maxSizeBytes: {type: integer, format: int64}
        expiresAt: {$ref: '#/components/schemas/LocalDateTime'}
...
    EventPage:
      type: object
      properties:
        records: {type: array, items: {$ref: '#/components/schemas/EventBriefVO'}}
        total: {type: integer, format: int64}
        size: {type: integer, format: int64}
        current: {type: integer, format: int64}
        pages: {type: integer, format: int64}
...
    ResultBase:
      type: object
      required: [code, data, msg]
      properties:
        code: {type: integer, format: int32, description: 成功为 200，失败为业务状态码}
        data: {}
        msg: {type: string}
    VoidResult:
      allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {type: 'null'}}}]
      example: {code: 200, data: null, msg: 成功}
    ErrorResult:
      allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {type: 'null'}}}]
...
    AuthorityUserListResult: {allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/AuthorityUserVO'}}}}]}
...
    EventIdResult: {allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {$ref: '#/components/schemas/EventIdVO'}}}]}
    TrackIdResult: {allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {$ref: '#/components/schemas/TrackIdVO'}}}]}
    EventInfoResult: {allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {$ref: '#/components/schemas/EventInfoVO'}}}]}
    EventPageResult: {allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {$ref: '#/components/schemas/EventPage'}}}]}
...
    UploadResult: {allOf: [{$ref: '#/components/schemas/ResultBase'}, {type: object, properties: {data: {$ref: '#/components/schemas/UploadVO'}}}]}
...